added tests to leaderboard, round, stage, tournament
