+++ backend/src/main/java/com/project/G1_T3/stage/service/StageServiceImpl.java
@@ -147,11 +147,6 @@ public class StageServiceImpl implements StageService {
         throw new IllegalArgumentException("There must be more than 1 player");
         }
 
-        // Validate input: Referees must not be null and should have at least 1 referee
-        if (stageDTO.getReferees() == null || stageDTO.getReferees().isEmpty()) {
-        throw new IllegalArgumentException("There must be at least 1 referee");
-        }
-
         // Create a new Stage entity and populate its fields
         Stage stage = new Stage();
         try {
@@ -164,8 +159,6 @@ public class StageServiceImpl implements StageService {
             stage.setStatus(stageDTO.getStatus() != null ? stageDTO.getStatus() : Status.SCHEDULED); // Default to
                                                                                                      // SCHEDULED
             stage.setTournament(tournament); // Set the associated tournament
-            // stage.setPlayers(stageDTO.getPlayers());
-            // stage.setReferees(stageDTO.getReferees());
 
             // Save the stage in the repository
             stage = stageRepository.save(stage);
